- Added internal expression cache. Cached expressions can be re-evaluated without compilation - Augmented the 're-evaluate' function - Renamed package
