+++ ANNOUNCE.rst
@@ -1,59 +1,15 @@
 ========================
-Announcing NumExpr 2.8.5
+NumExpr 2.8.5 (mod)
 ========================
 
-Hi everyone, 
-
-**Under development.**
-
-Project documentation is available at:
-
-http://numexpr.readthedocs.io/
-
-Changes from 2.8.4 to 2.8.5
+Changes from base 2.8.5
 ---------------------------
 
-**Under development.**
+In the modified version it is possible to precompile expressions and invoke them by name:
+  >>> a = np.array(1, 2, 3, 4, 5)
+  >>> b = np.array(6, 7, 8, 9, 0)
+  >>> ne.cache_expression("a + b", 'sum_ab')
+  >>> ne.re_evaluate('sum_ab')
+  array([ 0. ,  1.5,  3. ,  4.5,  6. ,  7.5,  9. , 10.5, 12. , 13.5])
 
-
-What's Numexpr?
----------------
-
-Numexpr is a fast numerical expression evaluator for NumPy.  With it,
-expressions that operate on arrays (like "3*a+4*b") are accelerated
-and use less memory than doing the same calculation in Python.
-
-It has multi-threaded capabilities, as well as support for Intel's
-MKL (Math Kernel Library), which allows an extremely fast evaluation
-of transcendental functions (sin, cos, tan, exp, log...) while
-squeezing the last drop of performance out of your multi-core
-processors.  Look here for a some benchmarks of numexpr using MKL:
-
-https://github.com/pydata/numexpr/wiki/NumexprMKL
-
-Its only dependency is NumPy (MKL is optional), so it works well as an
-easy-to-deploy, easy-to-use, computational engine for projects that
-don't want to adopt other solutions requiring more heavy dependencies.
-
-Where I can find Numexpr?
--------------------------
-
-The project is hosted at GitHub in:
-
-https://github.com/pydata/numexpr
-
-You can get the packages from PyPI as well (but not for RC releases):
-
-http://pypi.python.org/pypi/numexpr
-
-Documentation is hosted at:
-
-http://numexpr.readthedocs.io/en/latest/
-
-Share your experience
----------------------
-
-Let us know of any bugs, suggestions, gripes, kudos, etc. you may
-have.
-
-Enjoy data!
+It is required that variables remain the same type when cached expression is used.
