+++ MANIFEST.in
@@ -2,8 +2,8 @@ include MANIFEST.in
 
 include *.rst *.txt *.cfg site.cfg.example
 
-recursive-include numexpr *.cpp *.hpp *.py
-recursive-include numexpr/win32 *.c *.h
-exclude numexpr/__config__.py RELEASING.txt site.cfg
+recursive-include numexpr_mod *.cpp *.hpp *.py
+recursive-include numexpr_mod win32 *.c *.h
+exclude numexpr_mod __config__.py RELEASING.txt site.cfg
 
 recursive-include bench *.py
