+++ bench/poly.py
@@ -20,7 +20,7 @@ from __future__ import print_function
 import sys
 from time import time
 import numpy as np
-import numexpr as ne
+import numexpr_mod as ne
 
 
 #expr = ".25*x**3 + .75*x**2 - 1.5*x - 2"  # the polynomial to compute
