+++ numexpr_mod/cpuinfo.py
