+++ numexpr_mod/functions.hpp
