+++ numexpr_mod/missing_posix_functions.hpp
