+++ numexpr_mod/module.cpp
