+++ numexpr_mod/msvc_function_stubs.hpp
