+++ numexpr_mod/opcodes.hpp
