+++ numexpr_mod/str-two-way.hpp
