+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = numexpr
+name = numexpr_mod
 version = 2.8.5.dev1
 description = Fast numerical expression evaluator for NumPy
 author = David M. Cooke, Francesc Alted, and others
