+++ setup.py
@@ -18,7 +18,7 @@ import numpy as np
 with open('requirements.txt') as f:
     requirements = f.read().splitlines()
 
-with open('numexpr/version.py', 'w') as fh:
+with open('numexpr_mod/version.py', 'w') as fh:
     cfg = read_configuration("setup.cfg")
     fh.write('# THIS FILE IS GENERATED BY `SETUP.PY`\n')
     fh.write("version = '%s'\n" % cfg['metadata']['version'])
@@ -33,9 +33,9 @@ inc_dirs = [np.get_include()]
 libs = []  # Pre-built libraries ONLY, like python36.so
 clibs = []
 def_macros = []
-sources = ['numexpr/interpreter.cpp',
-           'numexpr/module.cpp',
-           'numexpr/numexpr_object.cpp']
+sources = ['numexpr_mod/interpreter.cpp',
+           'numexpr_mod/module.cpp',
+           'numexpr_mod/numexpr_object.cpp']
 extra_cflags = []
 extra_link_args = []
 
@@ -44,7 +44,7 @@ if platform.uname().system == 'Windows':
     if "MSC" in platform.python_compiler():
         extra_cflags = ['/O2']
     extra_link_args = []
-    sources.append('numexpr/win32/pthread.c')
+    sources.append('numexpr_mod/win32/pthread.c')
 else:
     extra_cflags = []
     extra_link_args = []
@@ -82,7 +82,7 @@ def setup_package():
     parse_site_cfg()
 
     numexpr_extension = Extension(
-        'numexpr.interpreter',
+        'numexpr_mod.interpreter',
         include_dirs=inc_dirs,
         define_macros=def_macros,
         sources=sources,
